+++ plugin/gdk/gvloadimage_gdk.c
@@ -37,7 +37,6 @@ typedef enum {
     FORMAT_ICO_CAIRO,
 } format_type;
 
-#ifdef HAVE_CAIRO_SURFACE_SET_MIME_DATA
 static void gdk_set_mimedata_from_file (cairo_surface_t *image, const char *mime_type, const char *file)
 {
     FILE *fp;
@@ -86,62 +85,34 @@ static void gdk_set_mimedata(cairo_surface_t *image, usershape_t *us)
     }
 }
 
-#endif /* HAVE_CAIRO_SURFACE_SET_MIME_DATA */
 static void gdk_freeimage(usershape_t *us)
 {
-#ifndef HAVE_CAIRO_SURFACE_SET_MIME_DATA
-    g_object_unref(us->data);
-#else /* HAVE_CAIRO_SURFACE_SET_MIME_DATA */
     cairo_surface_destroy(us->data);
-#endif /* HAVE_CAIRO_SURFACE_SET_MIME_DATA */
 }
 
-#ifndef HAVE_CAIRO_SURFACE_SET_MIME_DATA
-static GdkPixbuf* gdk_loadimage(GVJ_t * job, usershape_t *us)
-#else /* HAVE_CAIRO_SURFACE_SET_MIME_DATA */
 static cairo_surface_t* gdk_loadimage(GVJ_t * job, usershape_t *us)
-#endif /* HAVE_CAIRO_SURFACE_SET_MIME_DATA */
 {
-#ifdef HAVE_CAIRO_SURFACE_SET_MIME_DATA
     cairo_t *cr = job->context; /* target context */
-#endif /* HAVE_CAIRO_SURFACE_SET_MIME_DATA */
     GdkPixbuf *image = NULL;
-#ifdef HAVE_CAIRO_SURFACE_SET_MIME_DATA
     cairo_surface_t *cairo_image = NULL;
     cairo_pattern_t *pattern;
-#endif /* HAVE_CAIRO_SURFACE_SET_MIME_DATA */
 
     assert(job);
     assert(us);
     assert(us->name);
 
     if (us->data) {
-#ifndef HAVE_CAIRO_SURFACE_SET_MIME_DATA
-        if (us->datafree == gdk_freeimage)
-             image = us->data; /* use cached data */
-        else {
-             us->datafree(us);        /* free incompatible cache data */
-             us->datafree = NULL;
-             us->data = NULL;
-#else /* HAVE_CAIRO_SURFACE_SET_MIME_DATA */
         if (us->datafree == gdk_freeimage) {
 	    cairo_image = cairo_surface_reference(us->data); /* use cached data */
 	} else {
 	    us->datafree(us);        /* free incompatible cache data */
 	    us->datafree = NULL;
 	    us->data = NULL;
-#endif /* HAVE_CAIRO_SURFACE_SET_MIME_DATA */
         }
     }
-#ifndef HAVE_CAIRO_SURFACE_SET_MIME_DATA
-    if (!image) { /* read file into cache */
-	if (!gvusershape_file_access(us))
-	    return NULL;
-#else /* HAVE_CAIRO_SURFACE_SET_MIME_DATA */
     if (!cairo_image) { /* read file into cache */
         if (!gvusershape_file_access(us))
             return NULL;
-#endif /* HAVE_CAIRO_SURFACE_SET_MIME_DATA */
         switch (us->type) {
             case FT_PNG:
             case FT_JPEG:
@@ -154,13 +125,7 @@ static cairo_surface_t* gdk_loadimage(GVJ_t * job, usershape_t *us)
             default:
                 image = NULL;
         }
-#ifdef HAVE_CAIRO_SURFACE_SET_MIME_DATA
-
-#endif /* HAVE_CAIRO_SURFACE_SET_MIME_DATA */
         if (image) {
-#ifndef HAVE_CAIRO_SURFACE_SET_MIME_DATA
-            us->data = image;
-#else /* HAVE_CAIRO_SURFACE_SET_MIME_DATA */
             cairo_save (cr);
             gdk_cairo_set_source_pixbuf (cr, image, 0, 0);
             pattern = cairo_get_source (cr);
@@ -170,20 +135,11 @@ static cairo_surface_t* gdk_loadimage(GVJ_t * job, usershape_t *us)
             cairo_restore (cr);
             gdk_set_mimedata (cairo_image, us);
             us->data = cairo_surface_reference(cairo_image);
-#endif /* HAVE_CAIRO_SURFACE_SET_MIME_DATA */
             us->datafree = gdk_freeimage;
         }
-#ifndef HAVE_CAIRO_SURFACE_SET_MIME_DATA
-	gvusershape_file_release(us);
-#else /* HAVE_CAIRO_SURFACE_SET_MIME_DATA */
         gvusershape_file_release(us);
-#endif /* HAVE_CAIRO_SURFACE_SET_MIME_DATA */
     }
-#ifndef HAVE_CAIRO_SURFACE_SET_MIME_DATA
-    return image;
-#else /* HAVE_CAIRO_SURFACE_SET_MIME_DATA */
     return cairo_image;
-#endif /* HAVE_CAIRO_SURFACE_SET_MIME_DATA */
 }
 
 static void gdk_loadimage_cairo(GVJ_t * job, usershape_t *us, boxf b, bool filled)
@@ -191,27 +147,17 @@ static void gdk_loadimage_cairo(GVJ_t * job, usershape_t *us, boxf b, bool fille
     (void)filled;
 
     cairo_t *cr = job->context; /* target context */
-#ifndef HAVE_CAIRO_SURFACE_SET_MIME_DATA
-    GdkPixbuf *image;
-#else /* HAVE_CAIRO_SURFACE_SET_MIME_DATA */
     cairo_surface_t *image;
-#endif /* HAVE_CAIRO_SURFACE_SET_MIME_DATA */
 
     image = gdk_loadimage(job, us);
     if (image) {
         cairo_save(cr);
 	cairo_translate(cr, b.LL.x, -b.UR.y);
 	cairo_scale(cr, (b.UR.x - b.LL.x)/(us->w), (b.UR.y - b.LL.y)/(us->h)); 
-#ifndef HAVE_CAIRO_SURFACE_SET_MIME_DATA
-        gdk_cairo_set_source_pixbuf (cr, image, 0, 0);
-#else /* HAVE_CAIRO_SURFACE_SET_MIME_DATA */
         cairo_set_source_surface (cr, image, 0, 0);
-#endif /* HAVE_CAIRO_SURFACE_SET_MIME_DATA */
         cairo_paint (cr);
         cairo_restore(cr);
-#ifdef HAVE_CAIRO_SURFACE_SET_MIME_DATA
         cairo_surface_destroy (image);
-#endif /* HAVE_CAIRO_SURFACE_SET_MIME_DATA */
     }
 }