+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - The Autotools build system can now detect a MacPorts-installed libANN. #1854
+- Support for versions of Cairo prior to 1.10 has been removed.
 
 ### Fixed